requirements.txt
flask
pandas
python-pptx
